Extract PremisesActions prior to changing behaviour / refactoring
We will be changing the behaviour of the PremisesActions module
to return the following actions (`IdentityBarMenuItem`s) based
on Roles:

| Action             | Roles            | notes                                                   |
| ------------------ | ---------------- | ------------------------------------------------------- |
| withdraw placement | WORKFLOW_MANAGER | we'll switch to CRU_TEAM                                |
| create placement   | WORKFLOW_MANAGER | we'll switch to CRU_TEAM                                |
| view calendar      | MANAGER          | deprecated feature, will return improved in "v2 manage" |

temp: remove premisesActions from index
