+++ server/utils/premises/index.test.ts
@@ -3,7 +3,6 @@ import {
   bedOccupancyEntryFactory,
   bedOccupancyRangeFactory,
   extendedPremisesSummaryFactory,
-  premisesFactory,
   premisesSummaryFactory,
 } from '../../testutils/factories'
 import {
@@ -11,7 +10,6 @@ import {
   mapApiOccupancyEntryToUiOccupancyEntry,
   mapApiOccupancyToUiOccupancy,
   overcapacityMessage,
-  premisesActions,
   premisesTableRows,
   summaryListForPremises,
 } from '.'
@@ -20,7 +18,6 @@ import { textValue } from '../applications/helpers'
 import paths from '../../paths/manage'
 import { linkTo } from '../utils'
 import { DateFormats } from '../dateUtils'
-import userDetails from '../../testutils/factories/userDetails'
 
 jest.mock('../addOverbookingsToSchedule')
 
@@ -307,66 +304,6 @@ describe('premisesUtils', () => {
           },
         ],
       ])
-    })
-  })
-
-  describe('premisesActions', () => {
-    describe('for users with the role "workflow_manager"', () => {
-      const user = userDetails.build({ roles: ['workflow_manager'] })
-
-      it('returns all the actions', () => {
-        const premises = premisesFactory.build()
-
-        expect(premisesActions(user, premises)).toEqual([
-          {
-            text: 'View calendar',
-            classes: 'govuk-button--secondary',
-            href: paths.premises.calendar({ premisesId: premises.id }),
-          },
-          {
-            text: 'Create a placement',
-            classes: 'govuk-button--secondary',
-            href: paths.bookings.new({ premisesId: premises.id }),
-          },
-          {
-            text: 'Manage beds',
-            classes: 'govuk-button--secondary',
-            href: paths.premises.beds.index({ premisesId: premises.id }),
-          },
-        ])
-      })
-    })
-
-    describe('for users with the role "legacy_manager"', () => {
-      const user = userDetails.build({ roles: ['legacy_manager'] })
-
-      it('returns only "Manage beds"', () => {
-        const premises = premisesFactory.build()
-
-        expect(premisesActions(user, premises)).toEqual([
-          {
-            text: 'Manage beds',
-            classes: 'govuk-button--secondary',
-            href: paths.premises.beds.index({ premisesId: premises.id }),
-          },
-        ])
-      })
-    })
-
-    describe('for users with the role "future_manager"', () => {
-      const user = userDetails.build({ roles: ['future_manager'] })
-
-      it('returns only "Manage beds"', () => {
-        const premises = premisesFactory.build()
-
-        expect(premisesActions(user, premises)).toEqual([
-          {
-            text: 'Manage beds',
-            classes: 'govuk-button--secondary',
-            href: paths.premises.beds.index({ premisesId: premises.id }),
-          },
-        ])
-      })
     })
   })
 })
